# Documentation/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ECMA Events
  description: >-
    API da aplicação gerida pela companhia de marketing ECMA para a gestão da
    organização de eventos com a colaboração das suas companhias parceiras.
servers:
  - url: 'http://localhost:8080/'
...
paths:
  /users:
    get:
      tags:
        - Users
      description: Obter todos os utilizadores do sistema
      operationId: getUsers
      parameters:
        - name: criteria
          in: query
          description: Filtrar os utilizadores que satisfazem o critério
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Lista com a informação sobre os utilizadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Users
      description: Adicionar um utilizador ao sistema
      operationId: addUser
      requestBody:
        description: Todos os necessários dados do utilizador
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Utilizador adicionado com sucesso
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/users/{id}':
    get:
      tags:
        - Users
      description: 'Obter o utilizador identificado pelo {id}'
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: Identificador do utilizador a obter
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Informação do utilizador
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Utilizador não encontrado
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Users
      description: 'Atualizar o utilizador identificado pelo {id}'
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: Identificador do utilizador a remover
          required: true
          schema:
            type: integer
            format: int64
        - name: user
          in: query
          description: Objeto utilizador que é atualizado
          required: true
          schema:
            $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Utilizador atualizado
        '400':
          description: Id especificado inválido
        '404':
          description: Utilizador não encontrado
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Users
      description: 'Remover do sistema o utilizador identificado pelo {id}'
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: Identificador do utilizador a remover
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Utilizador removido
        '404':
          description: Utilizador não encontrado
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/users/{id}/approverInProposals':
    get:
      tags:
        - Users
      description: 'Obter a lista de propostas que o utilizador {id} pode aprovar.'
      operationId: getUserApproverInProposals
      parameters:
        - name: id
          in: path
          description: Identificador do utilizador a obter
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Ids das propostas que o utilizador pode aprovar
        '404':
          description: Utilizador não encontrado
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Users
      description: 'Adicionar uma proposta à lista de aprovação do utilizador.'
      operationId: addUserApproverInProposals
      parameters:
        - name: id
          in: path
          description: Identificador do utilizador
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Todos os necessários dados da proposal
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Proposal'
      responses:
        '200':
          description: Id da proposta adicionado à lista de aprovação do utilizador
        '400':
          description: Id especificado do utilizador ou proposta inválidos
        '404':
          description: Utilizador ou proposta não encontrados
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/users/{uid}/approverInProposals/{pid}':
    delete:
      tags:
        - Users
      description: 'Remover a proposta {pid} da lista de aprovação do utilizador {uid}'
      operationId: deleteUserApproverInProposals
      parameters:
        - name: uid
          in: path
          description: Identificador do utilizador
          required: true
          schema:
            type: integer
            format: int64
        - name: pid
          in: path
          description: >-
            Identificador da proposta a remover da lista de aprovação do
            utilizador
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Proposta removida da lista de aporvação do utilizador
        '404':
          description: Utilizador ou porposta não encontrados
        default:
          description: Erro não esperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      required:
        - firstName
        - lastName
        - userName
        - emailAddress
        - passwordHash
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        firstName:
          type: string
          example: john
        lastName:
          type: string
          example: smith
        userName:
          type: string
          example: user1
        emailAddress:
          type: string
          example: user1@ecma.com
        passwordHash:
          type: string
          example: $2y$12$95z6mYtPI371qXLCTWzP0elvLNv7Znh8dlwoseWncIm2zQb4u94xS
        approverInProposals:
          type: array
          items:
            type: number
...
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string